Remove non-typing circularity
