+++ app/reducers/commonActions.ts
@@ -1,0 +1,3 @@
+export interface PuzzleAction {
+  type: string;
+}
