+++ app/reducers/puzzleReducer.ts
@@ -1,11 +1,10 @@
-import { Position, Direction, BLOCK, Symmetry, CheatUnit } from '../lib/types';
+import { Direction, CheatUnit } from '../lib/types';
 import { PlayWithoutUserT } from '../lib/dbtypes';
-import { CluedGrid, gridWithNewChar, nextNonBlock } from '../lib/viewableGrid';
-import { cellIndex, valAt, entryAtPosition } from '../lib/gridBase';
+import { CluedGrid, nextNonBlock } from '../lib/viewableGrid';
 import { AccountPrefsFlagsT } from '../lib/prefs';
-import { checkGrid } from '../lib/utils';
 import { PuzzleAction } from './commonActions';
 import { GridInterfaceState, gridInterfaceReducer } from './gridReducer';
+import { cheat, getCurrentTime } from './puzzleUtils';
 
 export interface PuzzleState extends GridInterfaceState {
   type: 'puzzle';
@@ -40,10 +39,6 @@ export interface PuzzleState extends GridInterfaceState {
   contestSubmitTime?: number;
   contestDisplayName?: string;
   contestEmail?: string;
-}
-
-export function isPuzzleState(state: GridInterfaceState): state is PuzzleState {
-  return state.type === 'puzzle';
 }
 
 export interface StopDownsOnlyAction extends PuzzleAction {
@@ -130,120 +125,6 @@ export interface LoadPlayAction extends PuzzleAction {
 }
 function isLoadPlayAction(action: PuzzleAction): action is LoadPlayAction {
   return action.type === 'LOADPLAY';
-}
-
-function cheatCells(
-  elapsed: number,
-  state: PuzzleState,
-  cellsToCheck: Position[],
-  isReveal: boolean
-) {
-  const revealedCells = new Set(state.revealedCells);
-  const verifiedCells = new Set(state.verifiedCells);
-  const wrongCells = new Set(state.wrongCells);
-  let grid = state.grid;
-
-  for (const cell of cellsToCheck) {
-    const ci = cellIndex(state.grid, cell);
-    const shouldBe = state.answers[ci];
-    if (shouldBe === undefined) {
-      throw new Error('oob');
-    }
-    if (shouldBe === BLOCK) {
-      continue;
-    }
-    const currentVal = valAt(state.grid, cell);
-    if (shouldBe === currentVal) {
-      verifiedCells.add(ci);
-    } else if (isReveal) {
-      revealedCells.add(ci);
-      wrongCells.delete(ci);
-      verifiedCells.add(ci);
-      grid = gridWithNewChar(grid, cell, shouldBe, Symmetry.None);
-      state.cellsUpdatedAt[ci] = elapsed;
-      state.cellsIterationCount[ci] += 1;
-    } else if (currentVal.trim()) {
-      wrongCells.add(ci);
-      state.cellsEverMarkedWrong.add(ci);
-    }
-  }
-  return checkComplete({
-    ...state,
-    grid,
-    wrongCells,
-    revealedCells,
-    verifiedCells,
-  });
-}
-
-function cheat(state: PuzzleState, cheatUnit: CheatUnit, isReveal: boolean) {
-  const elapsed = getCurrentTime(state);
-  let cellsToCheck: Position[] = [];
-  if (cheatUnit === CheatUnit.Square) {
-    cellsToCheck = [state.active];
-  } else if (cheatUnit === CheatUnit.Entry) {
-    const entry = entryAtPosition(state.grid, state.active)[0];
-    if (!entry) {
-      //block?
-      return state;
-    }
-    cellsToCheck = entry.cells;
-  } else {
-    // Puzzle
-    for (let rowidx = 0; rowidx < state.grid.height; rowidx += 1) {
-      for (let colidx = 0; colidx < state.grid.width; colidx += 1) {
-        cellsToCheck.push({ row: rowidx, col: colidx });
-      }
-    }
-  }
-  const newState = cheatCells(elapsed, state, cellsToCheck, isReveal);
-  return { ...newState, didCheat: true };
-}
-
-function checkComplete(state: PuzzleState) {
-  const [filled, success] = checkGrid(
-    state.grid.cells,
-    state.answers,
-    state.alternateSolutions
-  );
-  if (filled !== state.filled || success !== state.success) {
-    let currentTimeWindowStart = state.currentTimeWindowStart;
-    let dismissedKeepTrying = state.dismissedKeepTrying;
-    let bankedSeconds = state.bankedSeconds;
-    // Pause if success or newly filled
-    if (currentTimeWindowStart && (success || (filled && !state.filled))) {
-      bankedSeconds = getCurrentTime(state);
-      currentTimeWindowStart = 0;
-      dismissedKeepTrying = false;
-    }
-    return {
-      ...state,
-      filled,
-      success,
-      bankedSeconds,
-      currentTimeWindowStart,
-      dismissedKeepTrying,
-    };
-  }
-  return state;
-}
-
-export function postEdit(state: PuzzleState, cellIndex: number): PuzzleState {
-  state.wrongCells.delete(cellIndex);
-  if (state.autocheck) {
-    return checkComplete(cheat(state, CheatUnit.Square, false));
-  }
-  return checkComplete(state);
-}
-
-function getCurrentTime(state: PuzzleState) {
-  if (state.currentTimeWindowStart === 0) {
-    return state.bankedSeconds;
-  }
-  return (
-    state.bankedSeconds +
-    (new Date().getTime() - state.currentTimeWindowStart) / 1000
-  );
 }
 
 export function puzzleReducer(
